+++ plugins/main/public/components/common/data-grid/use-data-grid.ts
@@ -236,5 +236,6 @@ export const useDataGrid = (props: tDataGridProps): EuiDataGridProps => {
       onChangeItemsPerPage: onChangeItemsPerPage,
       onChangePage: onChangePage,
     },
+    setPagination,
   } as EuiDataGridProps;
 };