+++ plugins/main/public/components/common/search-bar/search-bar.tsx
@@ -13,10 +13,14 @@ export interface WzSearchBarProps extends SearchBarProps {
   preQueryBar?: React.ReactElement;
   postFilters?: React.ReactElement;
   hideFixedFilters?: boolean;
+  showSaveQueryButton?: boolean;
+  showSaveQuery?: boolean;
 }
 
 export const WzSearchBar = ({
   fixedFilters = [],
+  showSaveQueryButton = true,
+  showSaveQuery = true,
   preQueryBar,
   hideFixedFilters,
   postFilters,
@@ -48,7 +52,11 @@ export const WzSearchBar = ({
         >
           {preQueryBar ? <EuiFlexItem>{preQueryBar}</EuiFlexItem> : null}
           <EuiFlexItem grow={!preQueryBar}>
-            <SearchBar {...restProps} />
+            <SearchBar
+              {...restProps}
+              showFilterBar={showSaveQueryButton}
+              showSaveQuery={showSaveQuery}
+            />
           </EuiFlexItem>
         </EuiFlexGroup>
       ) : null}