+++ plugins/main/public/components/overview/amazon-web-services/dashboards/dashboard.tsx
@@ -48,13 +48,13 @@ const DashboardAWSComponents: React.FC = ({}) => {
 
   const [results, setResults] = useState<SearchResponse>({} as SearchResponse);
 
-  const { searchBarProps } = useSearchBar({
+  const { searchBarProps, fingerprint, autoRefreshFingerprint } = useSearchBar({
     indexPattern: dataSource?.indexPattern as IndexPattern,
     filters,
     setFilters,
   });
 
-  const { query, absoluteDateRange } = searchBarProps;
+  const { query, dateRangeFrom, dateRangeTo } = searchBarProps;
 
   useReportingCommunicateSearchContext({
     isSearching: isDataSourceLoading,
@@ -62,7 +62,7 @@ const DashboardAWSComponents: React.FC = ({}) => {
     indexPattern: dataSource?.indexPattern,
     filters: fetchFilters,
     query: query,
-    time: absoluteDateRange,
+    time: { from: dateRangeFrom, to: dateRangeTo },
   });
 
   useEffect(() => {
@@ -71,7 +71,7 @@ const DashboardAWSComponents: React.FC = ({}) => {
     }
     fetchData({
       query,
-      dateRange: absoluteDateRange,
+      dateRange: { from: dateRangeFrom, to: dateRangeTo },
     })
       .then(results => setResults(results))
       .catch(error => {
@@ -84,7 +84,10 @@ const DashboardAWSComponents: React.FC = ({}) => {
   }, [
     JSON.stringify(fetchFilters),
     JSON.stringify(query),
-    JSON.stringify(absoluteDateRange),
+    dateRangeFrom,
+    dateRangeTo,
+    fingerprint,
+    autoRefreshFingerprint,
   ]);
   return (
     <>
@@ -121,7 +124,7 @@ const DashboardAWSComponents: React.FC = ({}) => {
                     filters: fetchFilters || [],
                     useMargins: true,
                     id: 'aws-dashboard-tab',
-                    timeRange: absoluteDateRange,
+                    timeRange: { from: dateRangeFrom, to: dateRangeTo },
                     title: 'AWS dashboard',
                     description: 'Dashboard of the AWS',
                     query: query,
@@ -130,6 +133,7 @@ const DashboardAWSComponents: React.FC = ({}) => {
                       value: 15,
                     },
                     hidePanelTitles: false,
+                    lastReloadRequestTime: fingerprint,
                   }}
                 />
               </div>