+++ plugins/main/public/components/overview/github/panel/config/drilldown-organization.tsx
@@ -131,7 +131,10 @@ export const DrilldownConfigOrganization = (
                       filters: fetchFilters ?? [],
                       useMargins: true,
                       id: 'github-drilldown-action-dashboard-tab',
-                      timeRange: searchBarProps?.absoluteDateRange,
+                      timeRange: {
+                        from: searchBarProps.dateRangeFrom,
+                        to: searchBarProps.dateRangeTo,
+                      },
                       title: 'GitHub drilldown action dashboard',
                       description: 'Dashboard of the GitHub drilldown action',
                       query: searchBarProps.query,