+++ plugins/main/public/components/overview/virustotal/dashboard/dashboard.tsx
@@ -48,12 +48,12 @@ const DashboardVT: React.FC = () => {
 
   const [results, setResults] = useState<SearchResponse>({} as SearchResponse);
 
-  const { searchBarProps } = useSearchBar({
+  const { searchBarProps, fingerprint, autoRefreshFingerprint } = useSearchBar({
     indexPattern: dataSource?.indexPattern as IndexPattern,
     filters,
     setFilters,
   });
-  const { query, absoluteDateRange } = searchBarProps;
+  const { query, dateRangeFrom, dateRangeTo } = searchBarProps;
 
   useReportingCommunicateSearchContext({
     isSearching: isDataSourceLoading,
@@ -61,7 +61,7 @@ const DashboardVT: React.FC = () => {
     indexPattern: dataSource?.indexPattern,
     filters: fetchFilters,
     query: query,
-    time: absoluteDateRange,
+    time: { from: dateRangeFrom, to: dateRangeTo },
   });
 
   useEffect(() => {
@@ -70,7 +70,7 @@ const DashboardVT: React.FC = () => {
     }
     fetchData({
       query,
-      dateRange: absoluteDateRange,
+      dateRange: { from: dateRangeFrom, to: dateRangeTo },
     })
       .then(results => {
         setResults(results);
@@ -86,7 +86,10 @@ const DashboardVT: React.FC = () => {
     isDataSourceLoading,
     JSON.stringify(fetchFilters),
     JSON.stringify(query),
-    JSON.stringify(absoluteDateRange),
+    dateRangeFrom,
+    dateRangeTo,
+    fingerprint,
+    autoRefreshFingerprint,
   ]);
 
   return (
@@ -125,7 +128,7 @@ const DashboardVT: React.FC = () => {
                 filters: fetchFilters ?? [],
                 useMargins: true,
                 id: 'kpis-virustotal-dashboard-tab',
-                timeRange: absoluteDateRange,
+                timeRange: { from: dateRangeFrom, to: dateRangeTo },
                 title: 'KPIs Virustotal dashboard',
                 description: 'KPIs Dashboard of the Virustotal',
                 query: query,
@@ -134,6 +137,7 @@ const DashboardVT: React.FC = () => {
                   value: 15,
                 },
                 hidePanelTitles: true,
+                lastReloadRequestTime: fingerprint,
               }}
             />
             <DashboardByRenderer
@@ -147,7 +151,7 @@ const DashboardVT: React.FC = () => {
                 filters: fetchFilters ?? [],
                 useMargins: true,
                 id: 'virustotal-dashboard-tab',
-                timeRange: absoluteDateRange,
+                timeRange: { from: dateRangeFrom, to: dateRangeTo },
                 title: 'Virustotal dashboard',
                 description: 'Dashboard of the Virustotal',
                 query: query,
@@ -156,6 +160,7 @@ const DashboardVT: React.FC = () => {
                   value: 15,
                 },
                 hidePanelTitles: false,
+                lastReloadRequestTime: fingerprint,
               }}
             />
           </div>